Split some PurchaseDetails fields into LegacyPurchaseDetails.
